requirements.txt
numpy
scipy
pandas
scikit-learn
pyyaml
matplotlib
umap-learn
joblib
